requirements.txt

customtkinter
pillow
stegano
oletools
python-magic
pygame
